feat: use Prisma validator for type-safe document queries
- Add Prisma.validator to define query shape once (Document.types.ts)
- Generate types automatically from validator using Prisma.DocumentVersionGetPayload
- Replace manual type definitions with generated types in DocumentsResults
- Eliminate code duplication and ensure type safety
- Single source of truth for complex Prisma queries

This ensures that changes to the query structure automatically update the types,
preventing type mismatches and reducing maintenance burden.
